# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personal Data Vault API",
        "description": "A secure API for managing personal data with encryption",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/personal-data": {
            "get": {
                "tags": [
                    "Personal Data"
                ],
                "summary": "Get all personal data for authenticated user",
                "description": "Display a listing of personal data",
                "operationId": "7e4eed18a2ed3bfe243758b98750cced",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "favorites",
                        "in": "query",
                        "description": "Filter by favorites only",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title, description, or tags",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Personal Data"
                ],
                "summary": "Create new personal data",
                "description": "Store a newly created personal data",
                "operationId": "0fb420bcfa4f04e65f9007c4b70de1b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "data_type",
                                    "data"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the data",
                                        "type": "string",
                                        "example": "Gmail Password"
                                    },
                                    "description": {
                                        "description": "Optional description",
                                        "type": "string",
                                        "example": "My Gmail account password"
                                    },
                                    "data_type": {
                                        "description": "Type of data",
                                        "type": "string",
                                        "enum": [
                                            "password",
                                            "note",
                                            "card",
                                            "account",
                                            "other"
                                        ],
                                        "example": "password"
                                    },
                                    "data": {
                                        "description": "The actual data (will be encrypted)",
                                        "type": "string",
                                        "example": "mySecurePassword123!"
                                    },
                                    "tags": {
                                        "description": "Tags for categorization",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "email",
                                            "google"
                                        ]
                                    },
                                    "is_favorite": {
                                        "description": "Mark as favorite",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personal data created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Personal data created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
        "/api/personal-data/search/{query}": {
            "get": {
                "tags": [
                    "Personal Data"
                ],
                "summary": "Search personal data",
                "description": "Search personal data",
                "operationId": "e7dc74e1e0cabeb452b7f42a81013319",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/personal-data/category/{category}": {
            "get": {
                "tags": [
                    "Personal Data"
                ],
                "summary": "Get personal data by category name",
                "description": "Get personal data by category",
                "operationId": "0cf86c2b3715e2dc48e3aa1d2c7b0084",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal data by category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/personal-data/export": {
            "get": {
                "tags": [
                    "Personal Data"
                ],
                "summary": "Export all personal data",
                "description": "Export personal data",
                "operationId": "3043dd69f4e9d015118f6f95d906cd7d",
                "responses": {
                    "200": {
                        "description": "Personal data exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "exported_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}